restructured download package
The file containing `CombinedProgressWatcher` is now named after the
class (`combined_progress_watcher.py`).
Helper package was removed, `CombinedProgressWatcher` is now directly
exposed on `minigalaxy.download.CombinedProgressWatcher`
