+++ minigalaxy/download/__init__.py
@@ -2,3 +2,4 @@
 
 # Flake8 thinks these imports are unused
 from minigalaxy.download.model import Download, DownloadType  # noqa: F401
+from minigalaxy.download.combined_progress_watcher import CombinedProgressWatcher  # noqa: F401
